feat: Observable theme loader
BREAKING CHANGE: The ThemeLoader is now expected to return an
Observable. When initializing the ThemeService, the ThemeLoader is
called once and the resulting Observable is subscribed to.

BREAKING CHANGE: Because they are generally no longer useful and hard to
retrofit, the ThemeService.savedTheme properties were removed.

The default ThemeLoader now watches for changes to the localStorage,
enabling the instant update of the theme when a different application
instance changes it, e.g. in a separate tab.

In addition, the ThemeService no longer auto-detects the theme and
always starts with the undefined theme.